pyproject.toml
[tool.poetry]
name = "yaml-indent"
version = "0.1.0"
description = ""
authors = ["Knut Olav Bøhmer <bohmer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
yaml-indent = "yaml_indent.yaml_indent:main"

